+++ src/0main/config_commands.h
@@ -2,7 +2,7 @@
 //////// command //////////
 struct struc_commandParts {
     bool isValidCommand;
-    size_t  huis;
+    uint8_t  huis;
     String  cmd;
     String arg1;
     String arg2;
@@ -11,33 +11,33 @@ struct struc_commandParts {
 struc_commandParts commandParts;
 
 //huis
-size_t huisMin = 1;
-size_t huisMax = 20;
+uint8_t huisMin = 1;
+uint8_t huisMax = 20;
 
 struct struc_huisRequest {
-    size_t huisnr;
-    size_t animnr;
-    size_t veld;
-    size_t lednr;
-    size_t rgbnr;
+    uint8_t huisnr;
+    uint8_t animnr;
+    uint8_t veld;
+    uint8_t lednr;
+    uint8_t rgbnr;
     String waarde;
 };
 struc_huisRequest huisRequest;
 
 // led
 struct struc_ledRequest {
-    size_t lednr;
-    size_t animnr;
-    size_t veld;
+    uint8_t lednr;
+    uint8_t animnr;
+    uint8_t veld;
     String waarde;
 };
 struc_ledRequest ledRequest;
 
 // rgb
 struct struc_rgbRequest {
-    size_t rgbnr;
-    size_t animnr;
-    size_t veld;
+    uint8_t rgbnr;
+    uint8_t animnr;
+    uint8_t veld;
     String waarde;
 };
 struc_rgbRequest rgbRequest;
