+++ src/0main/loadAnimatie.h
@@ -3,7 +3,7 @@ void loadAnimatie() {
   // Ini files zijn ingelezen.
   
   // leds
-  for (size_t lednr = 0; lednr < LED_NR_ITEMS; lednr++)
+  for (uint8_t lednr = 0; lednr < LED_NR_ITEMS; lednr++)
   {
     ledRequest.lednr  = lednr;
     ledRequest.animnr = huisSetting.animdefault;
@@ -11,7 +11,7 @@ void loadAnimatie() {
   }
   
   // rgb
-  for (size_t rgbnr = 0; rgbnr < RGB_NR_ITEMS; rgbnr++)
+  for (uint8_t rgbnr = 0; rgbnr < RGB_NR_ITEMS; rgbnr++)
   {
     rgbRequest.rgbnr  = rgbnr;
     rgbRequest.animnr = huisSetting.animdefault;
