+++ src/commands/cC00Help.h
@@ -1,6 +1,6 @@
 void cC00Help(){
     // Show Commands
-    for (size_t i = 0; i < 30 ; i++) {
+    for (uint8_t i = 0; i < 30 ; i++) {
         Print(validCommands[i].cmd + " ");
         Print(validCommands[i].arguments + " ");
         Println(validCommands[i].desc);
