+++ src/commands/cR05rgbHue.h
@@ -1,8 +1,8 @@
 void cR05rgbHue(){
     // Set HUE for rgb
-    size_t rgbnr = commandParts.arg1.toInt();
-    size_t nr = commandParts.arg2.toInt();
-    size_t hue = commandParts.arg3.toInt();
+    uint8_t rgbnr = commandParts.arg1.toInt();
+    uint8_t nr = commandParts.arg2.toInt();
+    uint8_t hue = commandParts.arg3.toInt();
     if (rgbnr >= 0 && rgbnr < RGB_NR_ITEMS) {
         switch (nr) {
             case 1:
