+++ src/led/ledLoadAnim.h
@@ -1,7 +1,7 @@
 #pragma once
 String ledLoadAnim() {
   // Ini files zijn ingelezen, nog eens toepassen wat daar staat
-  for (size_t lednr = 0; lednr < LED_NR_ITEMS; lednr++)
+  for (uint8_t lednr = 0; lednr < LED_NR_ITEMS; lednr++)
   {
     ledRequest.lednr  = lednr;
     ledRequest.animnr = huisSetting.animdefault;
