+++ src/rgb/rgbValueGet.h
@@ -1,7 +1,7 @@
 String rgbValueGet(){
-    size_t rgbnr = rgbRequest.rgbnr;
-    size_t animnr = rgbRequest.animnr;
-    size_t veldnr = rgbRequest.veld;
+    uint8_t rgbnr = rgbRequest.rgbnr;
+    uint8_t animnr = rgbRequest.animnr;
+    uint8_t veldnr = rgbRequest.veld;
     String answertoSend = "";
     switch (veldnr)
       {
